fixed some more decoding issues

+++ src/qrisp/qtypes/quantum_float.py
@@ -311,7 +311,8 @@ class QuantumFloat(QuantumVariable):
         if self.signed:
             res = signed_int_iso_inv(i, self.size - 1) * 2.0**self.exponent
         else:
-            res = i * 2**self.exponent
+            from jax.numpy import float32
+            res = i * float32(2)**self.exponent
 
         if self.exponent >= 0:
             if isinstance(res, (int, float)):
